micropython-button: groundwork
